+++ tesseract_environment/src/core/environment.cpp
@@ -332,6 +332,12 @@ EnvState::ConstPtr Environment::getCurrentState() const
   return current_state_;
 }
 
+std::chrono::high_resolution_clock::duration Environment::getCurrentStateTimestamp() const
+{
+  std::shared_lock<std::shared_mutex> lock(mutex_);
+  return current_state_timestamp_;
+}
+
 tesseract_scene_graph::Link::ConstPtr Environment::getLink(const std::string& name) const
 {
   std::shared_lock<std::shared_mutex> lock(mutex_);
@@ -676,6 +682,7 @@ void Environment::getCollisionObject(tesseract_collision::CollisionShapesConst& 
 
 void Environment::currentStateChanged()
 {
+  current_state_timestamp_ = std::chrono::high_resolution_clock::now().time_since_epoch();
   current_state_ = std::make_shared<EnvState>(*(state_solver_->getCurrentState()));
   if (discrete_manager_ != nullptr)
     discrete_manager_->setCollisionObjectsTransform(current_state_->link_transforms);
@@ -934,18 +941,6 @@ bool Environment::applyCommandsHelper(const Commands& commands)
       {
         auto cmd = std::static_pointer_cast<const AddKinematicsInformationCommand>(command);
         success &= applyAddKinematicsInformationCommand(cmd);
-        break;
-      }
-      case tesseract_environment::CommandType::CHANGE_DEFAULT_CONTACT_MARGIN:
-      {
-        auto cmd = std::static_pointer_cast<const ChangeDefaultContactMarginCommand>(command);
-        success &= applyChangeDefaultContactMarginCommand(cmd);
-        break;
-      }
-      case tesseract_environment::CommandType::CHANGE_PAIR_CONTACT_MARGIN:
-      {
-        auto cmd = std::static_pointer_cast<const ChangePairContactMarginCommand>(command);
-        success &= applyChangePairContactMarginCommand(cmd);
         break;
       }
       case tesseract_environment::CommandType::CHANGE_COLLISION_MARGINS:
@@ -1562,39 +1557,6 @@ bool Environment::applyAddKinematicsInformationCommand(const AddKinematicsInform
 {
   if (!manipulator_manager_->addKinematicsInformation(cmd->getKinematicsInformation()))
     return false;
-
-  ++revision_;
-  commands_.push_back(cmd);
-
-  return true;
-}
-
-bool Environment::applyChangeDefaultContactMarginCommand(const ChangeDefaultContactMarginCommand::ConstPtr& cmd)
-{
-  collision_margin_data_.setDefaultCollisionMargin(cmd->getDefaultCollisionMargin());
-
-  if (continuous_manager_ != nullptr)
-    continuous_manager_->setDefaultCollisionMarginData(cmd->getDefaultCollisionMargin());
-
-  if (discrete_manager_ != nullptr)
-    discrete_manager_->setDefaultCollisionMarginData(cmd->getDefaultCollisionMargin());
-
-  ++revision_;
-  commands_.push_back(cmd);
-
-  return true;
-}
-
-bool Environment::applyChangePairContactMarginCommand(const ChangePairContactMarginCommand::ConstPtr& cmd)
-{
-  for (const auto& link_pair : cmd->getPairCollisionMarginData())
-    collision_margin_data_.setPairCollisionMargin(link_pair.first.first, link_pair.first.second, link_pair.second);
-
-  if (continuous_manager_ != nullptr)
-    continuous_manager_->setCollisionMarginData(collision_margin_data_);
-
-  if (discrete_manager_ != nullptr)
-    discrete_manager_->setCollisionMarginData(collision_margin_data_);
 
   ++revision_;
   commands_.push_back(cmd);