+++ tesseract_kinematics/include/tesseract_kinematics/kdl/kdl_inv_kin_chain_nr.h
@@ -89,6 +89,8 @@ public:
 
   void setLimits(tesseract_common::KinematicLimits limits) override;
 
+  std::vector<Eigen::Index> getRedundancyCapableJointIndices() const override;
+
   unsigned int numJoints() const override;
 
   const std::string& getBaseLinkName() const override;