+++ tesseract_kinematics/src/kdl/kdl_inv_kin_chain_nr.cpp
@@ -98,15 +98,6 @@ IKSolutions KDLInvKinChainNR::calcInvKinHelper(const Eigen::Isometry3d& pose,
   if (tesseract_common::satisfiesPositionLimits(solution, kdl_data_.limits.joint_limits))
     solution_set.push_back(solution);
 
-  // Add redundant solutions
-  IKSolutions redundant_sols = getRedundantSolutions<double>(solution, kdl_data_.limits.joint_limits);
-  if (!redundant_sols.empty())
-  {
-    solution_set.insert(end(solution_set),
-                        std::make_move_iterator(redundant_sols.begin()),
-                        std::make_move_iterator(redundant_sols.end()));
-  }
-
   return solution_set;
 }
 
@@ -180,6 +171,11 @@ void KDLInvKinChainNR::setLimits(tesseract_common::KinematicLimits limits)
     throw std::runtime_error("Kinematics limits assigned are invalid!");
 
   kdl_data_.limits = std::move(limits);
+}
+
+std::vector<Eigen::Index> KDLInvKinChainNR::getRedundancyCapableJointIndices() const
+{
+  return kdl_data_.redundancy_indices;
 }
 
 unsigned int KDLInvKinChainNR::numJoints() const { return kdl_data_.robot_chain.getNrOfJoints(); }