+++ tesseract_kinematics/src/ur/ur_inv_kin.cpp
@@ -251,15 +251,6 @@ IKSolutions URInvKin::calcInvKin(const Eigen::Isometry3d& pose, const Eigen::Ref
 
     // Add solution
     solution_set.push_back(eigen_sol);
-
-    // Add redundant solutions
-    IKSolutions redundant_sols = getRedundantSolutions<double>(eigen_sol, limits_.joint_limits);
-    if (!redundant_sols.empty())
-    {
-      solution_set.insert(end(solution_set),
-                          std::make_move_iterator(redundant_sols.begin()),
-                          std::make_move_iterator(redundant_sols.end()));
-    }
   }
 
   return solution_set;
@@ -317,6 +308,8 @@ void URInvKin::setLimits(tesseract_common::KinematicLimits limits)
 
   limits_ = std::move(limits);
 }
+
+std::vector<Eigen::Index> URInvKin::getRedundancyCapableJointIndices() const { return redundancy_indices_; }
 
 const std::string& URInvKin::getBaseLinkName() const { return base_link_name_; }
 const std::string& URInvKin::getTipLinkName() const { return tip_link_name_; }