+++ tesseract_srdf/CHANGELOG.rst
@@ -2,6 +2,11 @@
 Changelog for package tesseract_srdf
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
+0.5.0 (2021-07-02)
+------------------
+* Fix SRDF REP and ROP positioner joint parsing
+* Contributors: Levi Armstrong
+
 0.4.1 (2021-04-24)
 ------------------